Our paper from 2011
Personal computing devices like smartphones and tablets are taking over the market; with such devices becoming more ubiquitous, users expect more from their devices. Since the mobile device does not have enough processing power, one prevalent programming model among application developers is to adopt a client-server model such that the computationally heavy work is transferred to a server while the mobile device is configured as a client. We believe that such a programming model would hinder growth in the long run, to that end we propose the COFA system to solve this problem. COFA is a system that can automatically, based on network conditions, offload computation to a server. With COFA system in place, the application developer would have to target only the Android phones. Also users of the applications would instantly gain the advantages of the system with faster responses and better overall …